requirements.txt
# SmartCAPEX AI Requirements
# ==========================

# Core scientific computing
numpy>=1.22.0
pandas>=1.5.0
scipy>=1.8.0

# Machine learning and AI
scikit-learn>=1.3.0
tensorflow>=2.10.0
dep>=1.3.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# GUI framework
# tkinter is included with Python standard library

# Additional utilities
Pillow>=9.0.0
requests>=2.28.0

# Optional (for advanced features)
# plotly>=5.0.0  # For interactive charts
# openpyxl>=3.0.0  # For Excel export
# xlsxwriter>=3.0.0  # For Excel writing
